requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
setuptools==75.6.0
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise==6.5.0
cloudinary~=1.44.0